@@ -97,15 +97,10 @@ class StreamHttpClient {
   void close({bool force = false}) => httpClient.close(force: force);
 
   StreamChatNetworkError _parseError(DioException exception) {
-    StreamChatNetworkError error;
     // locally thrown dio error
-    if (exception is StreamChatDioError) {
-      error = exception.error;
-    } else {
-      // real network request dio error
-      error = StreamChatNetworkError.fromDioException(exception);
-    }
-    return error..stackTrace = exception.stackTrace;
+    if (exception is StreamChatDioError) return exception.error;
+    // real network request dio error
+    return StreamChatNetworkError.fromDioException(exception);
   }
 
   /// Handy method to make http GET request with error parsing.
