@@ -164,7 +164,8 @@ class StreamChatState extends State<StreamChat> {
 
   @override
   void didChangeDependencies() {
-    final currentLocale = Localizations.localeOf(context).toString();
+    final currentLocale =
+        Localizations.localeOf(context).toString().toLowerCase();
     final availableLocales = Jiffy.getSupportedLocales();
     if (availableLocales.contains(currentLocale)) {
       Jiffy.setLocale(currentLocale);
